perf: reduce CPU load on F051 and other low-end MCUs
The F051 at 48MHz was running out of CPU cycles at high eRPM with
bidirectional dshot. This removes the major cycle sinks:

Shared code (all MCUs):
- replace the divide-by-2000 duty cycle scaling in the 20khz routine
  with a Q16 reciprocal multiply, the scale factor is recomputed in the
  main loop only when tim1_arr changes (cortex-m0 has no divider, each
  soft division cost 50+ cycles at 20khz)
- replace the recursive map() calls for throttle to duty cycle in
  setInput with startup-computed Q16 slopes (exact straight line,
  the old binary interpolation deviated up to 5/2000 from linear)
- make the current filter window a compile time constant so the
  divide becomes a multiply sequence
- decode dshot frames in a single pass into one packed word instead of
  a 16-entry int array plus three re-reads (also frees 64B of RAM)
- only write NVIC priorities when the dshot priority scheme actually
  changes instead of every main loop iteration
- enable LTO so hot helpers inline across translation units, CAN
  builds keep it off (gcc 10 maybe-uninitialized false positives in
  canard/dsdl) as does E230 (flash too tight for -O3 LTO inlining)

F051 specific:
- run the commutation hot path from 0WS RAM instead of 1WS flash:
  comparator/commutation/20khz IRQ handlers, commutate, comStep,
  interruptRoutine, PeriodElapsedCallback, tenKhzRoutine via a new
  .ramfunc section copied by the existing startup data init
- inline getCompOutputLevel, it was an out of line call in the
  comparator ISR zero cross filter loop (up to 12 calls per interrupt)
- stop enabling the SysTick interrupt, its handler is empty

Fixes found by LTO type checking:
- g431: commutation_interval extern was uint16_t, actual is uint32_t
- a153: aTxBuffer extern was int8_t[18], actual is uint8_t[49]
- dshot.c: remove unused mismatched zero_crosses extern

All 253 targets build. Worst case F051: flash 92.3%, RAM keeps 2.1KB
headroom above the heap/stack reserve.

+++ Mcu/g431/Src/stm32g4xx_it.c
@@ -19,7 +19,7 @@ extern volatile char dshot_telemetry;
 extern volatile char armed;
 extern volatile char out_put;
 extern volatile uint8_t compute_dshot_flag;
-extern volatile uint16_t commutation_interval;
+extern volatile uint32_t commutation_interval;
 
 int interrupt = 0;
